+++ app/controllers/Surveys.java
@@ -1,0 +1,12 @@
+/**
+ * 
+ */
+package controllers;
+
+/**
+ * @author Morteza Ansarinia <ansarinia@me.com>
+ *
+ */
+public class Surveys {
+
+}